All player functions interacting with backend game object.
All player actions work with the back end game object and update state except stay function.

Game room can now go for multiple rounds.

known error: stay function is conflicting with AIPlayer. The original while loop used to determine the computers actions is not working with the pushing of values from back end to front end. There is either an infinite loop of calls and/or the computer's hand is not being calculated correctly.